requirements.txt
# For GPU-enabled PyTorch (if needed for LLM finetuning):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# Core dependencies
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.128.0
fonttools==4.61.1
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
idna==3.11
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
narwhals==2.15.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.3.0
protobuf==3.20.3
pyarrow==23.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.3.2
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart>=0.0.6
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.6.1
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
starlette==0.50.0
streamlit==1.42.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0

# OCR Service dependencies (Windows compatible versions)
paddleocr==2.8.1
shapely>=2.0.0
ultralytics==8.0.137
